Simplify and fix the should_emit checks
looks_perly was matching too many non-Perl words like "module" causing
irrelevant posts to be emitted. This tightens the match to only return
true if the text includes "perl" or "cpan".

it also simplifies the emit checks from a scoring system to a boolean.
If any check fails, the post will not be emitted.